Add authorization checks to admin endpoints.

+++ src/main/java/github/paz/awardportal/controller/AdminOfficeController.java
@@ -1,6 +1,7 @@
 package github.paz.awardportal.controller;
 import github.paz.awardportal.model.Office.Office;
 import github.paz.awardportal.repository.OfficeRepository;
+import github.paz.awardportal.util.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,8 +16,12 @@ public class AdminOfficeController {
     @Autowired
     private OfficeRepository officeRepository;
 
+    @Autowired
+    SecurityUtil securityUtil;
+
     @GetMapping("/list")
     public String listOffices(Model model) {
+        securityUtil.assertAdminUser();
         List<Office> offices = officeRepository.findAll();
         model.addAttribute("offices", offices);
         return "redirect:/admin";
@@ -25,6 +30,7 @@ public class AdminOfficeController {
 
     @GetMapping("/addForm")
     public String addForm(Model model) {
+        securityUtil.assertAdminUser();
         Office office = new Office();
         model.addAttribute("office", office);
         return "offices/create-office-form";
@@ -33,6 +39,7 @@ public class AdminOfficeController {
 
     @GetMapping("/updateForm")
     public String updateForm(@RequestParam("officeId") Long id, Model model) {
+        securityUtil.assertAdminUser();
         Office office = officeRepository.getOne(id);
         model.addAttribute("office", office);
         return "offices/update-office-form";
@@ -41,8 +48,7 @@ public class AdminOfficeController {
 
     @PostMapping("/create")
     public String save(@ModelAttribute("office") Office office) {
-
-
+        securityUtil.assertAdminUser();
         officeRepository.save(office);
         return "redirect:/offices/list";
     }
@@ -50,6 +56,7 @@ public class AdminOfficeController {
 
     @PostMapping("/update")
     public String update(@ModelAttribute("office") Office office) {
+        securityUtil.assertAdminUser();
         officeRepository.save(office);
         return "redirect:/offices/list";
     }
@@ -57,6 +64,7 @@ public class AdminOfficeController {
 
     @GetMapping("/delete")
     public String delete(@RequestParam("officeId") Long id) {
+        securityUtil.assertAdminUser();
         officeRepository.deleteById(id);
         return "redirect:/offices/list";
     }
